requirements.txt
pytest-playwright
playwright
python-dotenv
Faker